rename : wallet class move vo package
